requirements.txt
Django==2.0.3
django-webpack-loader==0.6.0
djangorestframework==3.7.7
django-cors-headers==2.2.0
mysqlclient==1.3.12